+++ test/spec/xform.spec.js
@@ -108,7 +108,7 @@ describe( 'XForm', () => {
         const xf = loadXForm( 'appearances.xml' );
         const result = validator.validate( xf );
         const resultOc = validator.validate( xf, { openclinica: true } );
-        const ISSUES = 12;
+        const ISSUES = 13;
 
         it( 'outputs warnings', () => {
             expect( result.warnings.length ).to.equal( ISSUES );
@@ -120,6 +120,7 @@ describe( 'XForm', () => {
             expect( arrContains( result.warnings, /"compact" for question "e"/i ) ).to.equal( true );
             expect( arrContains( result.warnings, /"compact-19" for question "f"/i ) ).to.equal( true );
             expect( arrContains( result.warnings, /"numbers" for question "g"/i ) ).to.equal( true );
+            expect( arrContains( result.warnings, /"horizontal-compact" for question "k" .+ deprecated.+"compact"/i ) ).to.equal( true );
             expect( arrContains( result.warnings, /"field-list" for question "two"/i ) ).to.equal( true );
         } );
 
@@ -131,8 +132,9 @@ describe( 'XForm', () => {
             expect( resultOc.warnings.length ).to.equal( 0 );
         } );
 
-        it( 'outputs errors with --oc flag', () => {
-            expect( resultOc.errors.length ).to.equal( ISSUES );
+        it( 'outputs errors with --oc flag and ignores "deprecated" usage', () => {
+            expect( arrContains( result.errors, /deprecated/ ) ).to.equal( false );
+            expect( resultOc.errors.length ).to.equal( ISSUES - 2 );
         } );
 
     } );